map done. fuck.

+++ src/crc.c
@@ -33,9 +33,8 @@ const unsigned short Crc16Table[256] = {
     0x8201, 0x42C0, 0x4380, 0x8341, 0x4100, 0x81C1, 0x8081, 0x4040
 };
 
-unsigned short Crc16(unsigned char * pcBlock, unsigned short len)
-{
-    unsigned short crc = 0xFFFF;
+unsigned int Crc16(unsigned char * pcBlock, unsigned short len){
+    unsigned int crc = 0xFFFF;
 
     while (len--)
         crc = (crc >> 8) ^ Crc16Table[(crc & 0xFF) ^ *pcBlock++];
